feat: add NovelInfo
